Minor test and docs improvements for main OpenTelemetry module (#17)
* Add basic tests to OpenTelemetry module

This commit gets the basic tests running. I created a blank
DistributedContext::Manager class because that was undefined and the
OpenTelemetry.distributed_context_manager test started to fail.

* Convert Tracer, Meter and DistributedContext Manager to classes

Using classes is usually easier. We can easily create two tracers in a
test and compare them and we don't have to worry that much about the
shared state.

* Fix OpenTelemetry module documentation

Adds basic docs to OpenTelemetry module methods. Also fixes the
OpenTelemetry module comments.