Fix merge; initialize with hash instead of kwargs
The merge implementation was not working as intialize was expecting
kwargs, yet the internal storage of labels is a hash with string
keys and string values. It's easiest if we accept a hash with
keys and values in the same format as the labels are stored. So,
I updated initialize to expect a hash with string keys and values.
Another option would be to keep kwargs in initialize, but convert
string keys for labels to symbols, then use ** when merging, but
that seems like more effort than it's worth.