Improve Resource (#24)
* Add test for Resource

* Fix merge; initialize with hash instead of kwargs

The merge implementation was not working as intialize was expecting
kwargs, yet the internal storage of labels is a hash with string
keys and string values. It's easiest if we accept a hash with
keys and values in the same format as the labels are stored. So,
I updated initialize to expect a hash with string keys and values.
Another option would be to keep kwargs in initialize, but convert
string keys for labels to symbols, then use ** when merging, but
that seems like more effort than it's worth.

* Update Resource#merge semantics to comply with specification

* Enforce that label keys and values are strings

* Make Resource.create factory method and make constructor private

By introducing the create factory method we can validate user
input and reserve the constructor for internal use with data
that has already been processed. This allows us to save a hash
allocation, and type checking when we merge two existing resources.

* Expose enumerator for labels, rather than the underlying frozen hash

This adds a layer of protection for the underlying labels data structure
while still allowing users the ability to iterate over them.

* Inline check_and_freeze_labels in create factory method