sig-feedback: remove name and version from config
Removes the name and version from the config as the spec defines
these as coming from the instrumentation/adapter and not from the
instrumented library.

Adds convenience methods for the name and version and provide
those on the `Faraday` adapter class.

Updates the tracer configuration to use the new name and version.

Updates the example to remove the name and version from the call.