Add `rest-client` instrumentation adapter (#172)
Add `rest-client` instrumentation adapter
based on the adapter for Faraday and `ddtrace` code for `rest-client`.

The `rest-client` gem uses `restclient` internally for file and folder
naming, so I used that style for most of the internal adapter code.
`rest-client` is used for names referring to the gem name itself.

Related to #67