Update `faraday` adapter
Update `faraday` adapter
to match changes in the `rest-client` adapter.

In the example code, the install call is replaced
to work with the structure changes from #175.

Support for disabling span reporting via a custom override
is no longer provided since there is not a need for it at this time.
This also removes the need to support injection of custom middleware.

In the tracer middleware, the HTTP method attribute value
is set to an uppercase string to be compliant with the specification.
The name of the span is set to "HTTP #{verb}"
to reduce the cardinality of the span names as described in the spec.
The span status is now set when the response is received.
The span propagation is now done directly.

Propagated headers are now tested.