Add support for automated resource detection (#263)
* Add support for automated resource detection

This adds support for populating a resource with with Telemetry data
as well as Google Cloud Platform environment metata data.

For #230

* Fix typo typo in correlation context module

* Extract resource detection into a gem

To avoid introducing third party dependencies to the SDK for optional
functionality resource detection has been extracted into its own gem.

* Add resource-detectors README

* Add default Telemetry SDK resource create method

The specification requires a Telemetry::SDK resource, this adds a
method for create it directly from the resource class.  This also
initializes it by default on the configurator.  The resource
setter on the configurator class will merge the default resource
with the one provided.

* Add comment to disable Style/AccessModifierDeclarations in resource class

* Adjust tracer to use instrumentation library

* Re-add tracer_tests for name and version

* Add method documentation, update comments

Add resource documentation for configurator.
Update comments for resource constants.
Simplify resource merging in auto detector.

* Update copyright year, revert incorrect comment

* Fix typo in gcp resource detector test

* Set tracer resource in initializer

* Add tracer_provider reference to tracer

The tracer carries a reference to the tracer_provider that initialized it to
begin supporting multiple tracer providers.