fix! remove child_count
child_count is not required or supported by OTLP, so we don't need to
record it anymore in SpanData or track it in Span (we weren't, but we
had an unnecessary instance variable).

This is a breaking change for SpanExporters that may have used
SpanData's child_count. None of the exporters in this repo used it,
and I doubt anyone else did either.