refactor: rails instrumentation tests
The Rails instrumentation tests were initializing an app instance per test run,
this resulted in leaky state between tests.  The tests have been refactored
to only initialize! the rails application once on requiring the test helper.

We lose some potential flexibility in testing how configs impact the initialize
flow for the app, but we don't have any configs yet so we can solve that when
we come to that point.