refactor: rails instrumentation tests... again
This approach is combination of the first two attempts.  We declare a global
rails app in the test helper and assign it to ::Rails.application. The
intention is to keep CI speeds reasonable by not repeatedly initializing new
apps for each test.  The other motivation to this approach is due to some
undocumented behaviour around initializing multiple apps, there appears to be
state leaking between multiple `initialize!` calls. Tests by default should
either reference ::Rails.application or DEFAULT_RAILS_APP.

Any subsequent apps initialized will not be available under ::Rails.application
and should only be used to test variations in configuration when needed.

Example use case is testing an application with an exceptions_app configured,
and one without.  This requires a new app to be initialized.