feat: instrument rails (#454)
* feat: instrument rails

* refactor: rails instrumentation tests

The Rails instrumentation tests were initializing an app instance per test run,
this resulted in leaky state between tests.  The tests have been refactored
to only initialize! the rails application once on requiring the test helper.

We lose some potential flexibility in testing how configs impact the initialize
flow for the app, but we don't have any configs yet so we can solve that when
we come to that point.

* refactor: rails instrumentation tests... again

This approach is combination of the first two attempts.  We declare a global
rails app in the test helper and assign it to ::Rails.application. The
intention is to keep CI speeds reasonable by not repeatedly initializing new
apps for each test.  The other motivation to this approach is due to some
undocumented behaviour around initializing multiple apps, there appears to be
state leaking between multiple `initialize!` calls. Tests by default should
either reference ::Rails.application or DEFAULT_RAILS_APP.

Any subsequent apps initialized will not be available under ::Rails.application
and should only be used to test variations in configuration when needed.

Example use case is testing an application with an exceptions_app configured,
and one without.  This requires a new app to be initialized.

* feat: Add minimal tracing example

This example is a minimal Rails application that demonstrates tracing
action controller requests.

* refactor: simplified rails test setup

There was a lot of excess duplication, which has been moved to a single config
module which will apply rails specific version configs as part of
the initialize_app call.

* fix: Exceptions controller now returns 500s

The example exceptions controller was a little confusing as it was returning
200 responses.

* fix: railtie executes before initialize

The railtie initialization order is very particular, if configure the
instrumentation from a railtie its much more forgiving.  Cannot
rely on that assumption so the railtie now fires before the initialize hook,
which is compatible with the conventional config/initializers pattern.

* refactor: use rack instrumentation gem

Make use of existing rack instrumentation and append additional information
to the rack span in action controller metal.

* fix: use context instead of thread locals

Replace the usage of thread locals with context for retrieving the current rack
span.

* Update instrumentation/rails/lib/opentelemetry/instrumentation/rails/patches/action_controller/metal.rb

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>

* fix: Add rails to releases list

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>