feat: Add instrumentation for postgresql (pg gem)
This commit adds tracing for postgresql, via the `pg` gem.

The following methods are traced:
- exec / query / sync_exec / async_exec
- exec_params / sync_exec_params / async_exec_params
- prepare / sync_prepare / async_prepare
- exec_prepared / sync_exec_prepared / async_exec_prepared

We instrument all of these methods via meta-programming, since otherwise
it'd be incredibly repetitive. The methods can actually be broken down
into three groups - "exec-ish", "prepare-ish", and "exec prepared-ish"
methods. Every method within a group takes the same arguments, and so
we can trace each method in a group in an identical fashion.

Note that the 'sync' and 'async' variants require no special handling,
because they are not async interfaces _per se_ - rather, they are
non-blocking. See the `pg` docs for more info: https://www.rubydoc.info/gems/pg/PG%2FConnection:exec

We've implemented SQL sanitization the same way that the Mysql2 gem did
it - which seems to be done by copying NewRelic. We should consider
extracting that from the NewRelic agent and making it a common shared
library here.

Possibly worth discussion is how we've handled the span name. We've
chosen the format `OPERATION_NAME DATABASE_NAME`, taking the first word
in the SQL string as the operation if it matches a known good list we
pulled from the Postgres docs. This is similar to how the python and
javascript postgres implementations behave, but is slightly different
than how the ruby mysql instrumentation does it. When falling back to
safe value, we also choose to simply use `DATABASE_NAME`, which seems to
be in line with current semantic conventions.

While tests pass and we think this is safe, it should be noted that we
haven't actually put this into production testing yet - there could be
horrible bugs still. :)

Fixes #523