feat: Add lru cache for prepared statement names
This uses an LRU cache to store the 50 most recently-prepared database
statements. The goal is to provide something more descriptive than
'EXECUTE a1', which is basically what you get when Rails aggressively
prepares and executes statements.

50 seems to me like a reasonable trade-off between usefulness and
capping memory growth. However, it is technically unbounded if you are
not sanitizing SQL - although we could decide to truncate unconditionally.