rescue missed otlp exporter network errors
The Net::HTTP client has a few more errors that need to be caught or the
exporter will crash if the upstream agent/collector is temporarily
unavailable.

You can simulate one of the SystemCallErrors, Errno::ECONNREFUSED by
 restarting a local OpenTelemetry collector while an app is trying to
export spans to it. Without this change the exporter will crash, but
with this change the exporter should recover when the collector becomes
available