feat!: Add simple 'recording' attr_accessor to InMemorySpanExporter (#711)
* feat: Add simple 'recording' attr_accessor to InMemorySpanExporter

When true, the InMemorySpanExporter will retain the spans passed to it,
for later inspection. When false, they will simply be discarded. We
default things to `true` for easy setup and experimentation - anyone
using this in a test suite should actually turn recording off and then
selectively enable it when needed.

* Update sdk/test/opentelemetry/sdk/trace/export/in_memory_span_exporter_test.rb

Co-authored-by: Robert <robertlaurin@users.noreply.github.com>

Co-authored-by: Robert <robertlaurin@users.noreply.github.com>