feat: Add ActiveJob instrumentation
This is a slightly modified version of the ActiveJob instrumentation
that we have been running for a few weeks now, on two different internal
applications. It is somewhat basic, but it works well enough.

There are a few things of note:
- We track a few custom attributes for ActiveJob - I'm not sure if the
  naming is good, however.
- We do not attempt to pass any kind of context propagation through the
  job system. I have successfully done that via job-argument-mangling,
  but didn't do so here. We've found that doing so leads to unwieldy
  traces for web requests that enqueue background jobs.
- I've tried to follow the overall messaging semantic conventions, since
  'job' semantic conventions have not materialized yet. I'll update this
  PR if they are finalized before this is merged; otherwise we can make
  those updates later.
- The test suite is a tad slow - to test a few features, I needed to
  explicitly use a real job system - in this case, the `async` adapter.
  It's all still in-process, but uses threads as a queueing system. I
  found the easiest thing to do was to just call `sleep` in a few places.

Probably the most important things to think about when reviewing are:
- Are the attributes added correct (or - are they at least close to
  correct until we have proper job semantic conventions)?
- Following a job produce -> job enqueue requires searching your tracing
  system on the value of `messaging.message_id`. Is that ergonomic or
  should we do something different?

I'd appreciate anyone who's willing to test this out in a production
system - I think. We're going to test this version internally as well,
since it differs in a few small ways from what we actually run today.