feat: option to obfuscate redis arguments
This change introduces an option for the Redis instrumentation that
will obfuscate all arguments in the db.statement attribute.

The purpose of this change is to give clients the option to hide
potentially sensitive information such as personally identifiable
information. This is similar to changes that have already been made to
the mysql and pg instrumentation.