Change option handling to make testing easier
The tests were failing depending on the run order; and it's basically
all my fault. Rather than re-install the instrumentation between tests
(which has its own downsides!) I was just reaching in and fiddling
around with the internal options. However, depending on run order the
instrumentation options could actually become unset.

This cleans up the test suite and tests passed 20 times in a row
(including with previously-failed randomization seeds) so I think this
is in okay shape.