fix!: Use 55681 as the default OLTP exporter port
In the interest of making this library easier to use by default, this PR
proposes a simple change: allow our OTLP exporter to default to using
port 55681, rather than 4317. Port 4317 is the collector's GRPC port,
and our exporter doesn't support GRPC. The default HTTP port is 55681,
and so we use it throughout.

However, this does seem to be against the specification in some ways.
The specification doesn't *require* that you support GRPC, however it
defines the default value for the environment variable as if you do.

I think this is acceptable; and we can default to port 4317 and GRPC
once we have support for it and are comfortable that it performs well.