refactor: Remove redundant methods in instrumentation/rack
While I do not know the precise provenance of these methods (nor do I
care to futz around with `git` any further to find out), the current API
contains similar methods that we can rely upon for this functionality.

Specifically, we remove:

- `OpenTelemetry::Instrumentation::Rack.context_with_span`
- `OpenTelemetry::Instrumentation::Rack.with_span`

`with_span` was being used, but could be replaced by the API method.
`current_span` is not being used _here_ but is being used by the `Rails`
instrumentation, so it stays.