Update docs to rely more on environment variable configuration
This PR updates our documentation in a few ways:
- Replaces some outdated configuration examples referencing a tracer factory
- Cleans up a reference to a non-existent LMDB example
- Sets the console span exporter for all current instrumentation examples
- Updates docs to reflect how to configure things via the environment whenever possible or reasonable.

Note that this means our examples may not flush to the console since the
BatchSpanProcessor is selected by default. I think this is okay,
honestly, but if it comes up we could figure out a way to select the
SimpleSpanProcessor (as some examples were doing) via the environment.

Fixes #872