feat: Add ActionView instrumentation via ActiveSupport::Notifications (#796)
* feat: Add ActionView instrumentation via ActiveSupport::Notifications

This commit adds a safe (I hope) subscription framework for
ActiveSupport::Notifications and then uses that framework to instrument
ActionView.

Key to this PR is the fact that Rails allows you to swap out the queue for
ActiveSupport::Notifications; although documentation on that is sparse.

We implement a thin wrapper around the existing (and default) Fanout
queue and a "SpanSubscriber" that handles the actual event instrumentation.
Together, they do three special things:

- Sort the subscribers for a notification so that our "SpanSubscriber"
  class always runs first.
- Start a span when the event starts
- Finish the span when the event finishes, merges in the payload as
  span attributes, and then toggles the error bit if there was an
  uncaught exception.

There are two main problems with ActiveSupport::Notifications as they
relate to telemetry, and we attempt to solve them both here:

1. Any *other* subscriber to the notification in question could throw,
   and that could mean we are left with dangling spans. Our solution to
   this lies in how we sort the subscribers for a notification when
   handling the events in the queue wrapper. By ensuring that our
   handlers run first, we ensure that anyone *else* who might throw
   doesn't affect our span handling.

2. Span stack handling can be a nightmare. Here, we rely on a
   undocumented (?) feature (??) of notifications: if the callable
   passed in with the subscription responds to `start`/`finish` methods,
   then Rails will call `start` and `finish` on your queue object. And
   the return value of `start` from your queue is saved around the
   execution of the instrumented block, and passed back in to the
   `finish` method on your queue object. We can use that and pass around
   the span that was started *and* the previous context to restore, and
   that greatly simplifies the state management.

There are, of course, tradeoffs:

1. In this implementation, we don't allow for any kind of real
   specialization in the "SpanSubscriber" object. We could in the
   future; and individual instrumentations could always subclass it.
2. You always lose some detail by instrumenting in this way - it's never
   going to be as rich as patching the classes directly; but it is
   something of a handy-dandy shortcut.
3. We add a dependency on the `rails` instrumentation. I don't know if
   this is the right thing to do here or not, and I don't know if
   potentially installing the instrumentation many times will have a
   huge problem.

This PR is not really done; I have tested it locally but not
extensively. And notably I haven't written any tests whatsoever for it.
But I'm putting it up now because I would like early feedback on this
approach!

* Remove unintentional .ruby-version commit ... again

* Update opentelemetry-instrumentation-all.gemspec

* Memoize span name, and tracer

This allows us to stop manipulating a new string on every span, as well
as sets the instrumentation name to `Instrumentation::ActionView` rather
than `Instrumentation::Rails`.

* Add span_subscriber tests

* Add tests for fanout notification queue

* Fixup tests for ActionView - not that there are many..

* Add example and flesh out readme a bit

* Note that rails instrumentation is required at present

* appease the cop

* silence the cop, harder this time

* Perhaps one last silencing attempt

* Update for new context / span error methods

* Reject invalid AS::Notifications payload values

According to the spec, we should allow:
- string
- boolean
- numeric
- arrays where all members are the same type (and that type is string,
  boolean, or numeric)
- empty arrays

We filter out anything that doesn't meet that criteria (and we continue
filtering out exception/exception_object attributes). We allow symbols
as values and stringify them to meet the spec (we do not transform
anything else).

* Support disallowing certain AS::Notification payload keys

These will never be set as a span attribute value.

* Allow AS:Notification payloads to be arbitrarily transformed

* Use local opentelemetry-semantic_conventions gem

* Update appraisal gemfiles for local semantic_conventions gem

* Swap notification queue class in before_initialize hook

The initializer approach was running into difficulties in anything other
than a very plain-vanilla rails app. This seems a little better.

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>