feat: add Que instrumentation (#918)
From the home page:
> [que][1] is a high-performance job queue that improves the reliability
of your application by protecting your jobs with the same ACID
guarantees as the rest of your data

It's similar to delayed_job and sidekiq.

There are two parts here:
* Patching `Que::Job.enqueue` which is used when enqueuing and new job.
* Adding a `job_middleware` which is used when processing the job.

Que doesn't have a good place to store metadata. There are
basically two options: the job payload and the job tags.

Using the job payload is very brittle. We'd have to modify
existing Hash arguments or add a new argument when there are
no arguments we can modify. If the server side is not using
this instrumentation yet (e.g. old jobs before the
instrumentation was added or when instrumentation is being
added to client side first) then the server can error out due
to unexpected payload.

The second option (which we are using here) is to use tags.
They also are not meant for tracing information but they are
much safer to use than modifying the payload.

[1]: https://github.com/que-rb/que

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>