Don't error on non-string/utf8 kafka message keys
The instrumentation for Kafka was assuming that message keys are only
valid utf8 strings, but message keys can be any string of bytes. See:
https://cwiki.apache.org/confluence/display/KAFKA/A+Guide+To+The+Kafka+Protocol#AGuideToTheKafkaProtocol-Responses

This was causing encoding errors in the exporters, and lost spans, as they
were trying to encode string into utf8 before transport.

This fix converts keys that aren't valid utf8 strings into hex strings.
This seems to be consistent with how other byte strings are encoded such
as with trace ids.