test set_attribute with a float value and length limit configured
This is a regression test. Before the change to check if an attribute
value is a string before checking its size when truncating the
value of an attribute the code would crash due to calling `.size`
on a float.

Strings, arrays and integers all implement `.size` so this omision
is only caught by the float type of attribute value.i