feat(active_job): Make the context available in ActiveJob v7 notifications
Instead of using a callback to propagate the context, we can override
`ActiveJob::Base#perform_now`, which is the entrypoint to execution
logic. It is called right after the job is deserialized.

This is the same pattern that is used in `ActiveJob::Instrumentation`
in v7. Instead of using a callback like it did in previous versions,
the instrumentation in v7 overrides `#perform_now`. See:
https://github.com/rails/rails/pull/40927/files#diff-41ecb8d7126485e207a879c9e3e28f73837309b8370d0db939a23874b97f66b1L12