fix(excon): Fix error with upper-case HTTP methods
When Excon was called with an HTTP method in upper-case, the
instrumentation failed to capture the 'http.method' and set it to `nil`.
This caused the response instrumentation to log the following error:
```
OpenTelemetry error: invalid span attribute value type NilClass for key 'http.method' on span 'HTTP '
```

This commit fixes the issue and captures the 'http.method' regardless of
its casing.