feat: Make the context available in ActiveJob notifications (#1145)
* feat(active_job): Make the context available in ActiveJob notifications

Previously, subscribers to 'perform.active_job' had no access to the
tracing context. Only the job and its callbacks had access to it.

This was a result of adding OpenTelemetry's instrumentation callback
after the notifications instrumentation that is configured in
`ActiveJob::Logging` (in ActiveJob 6.x).

This commit fixes the issue by prepending OpenTelemetry's instrumentation
callback before all others.

In ActiveJob v7, however, notifications instrumentation happens before
any callbacks are called, so the current solution doesn't work, sadly

* feat(active_job): Make the context available in ActiveJob v7 notifications

Instead of using a callback to propagate the context, we can override
`ActiveJob::Base#perform_now`, which is the entrypoint to execution
logic. It is called right after the job is deserialized.

This is the same pattern that is used in `ActiveJob::Instrumentation`
in v7. Instead of using a callback like it did in previous versions,
the instrumentation in v7 overrides `#perform_now`. See:
https://github.com/rails/rails/pull/40927/files#diff-41ecb8d7126485e207a879c9e3e28f73837309b8370d0db939a23874b97f66b1L12

Co-authored-by: Andrew Hayworth <ahayworth@gmail.com>
Co-authored-by: Robert <robertlaurin@users.noreply.github.com>