Bump semantic_conventions gem to 1.10.0 (#1171)
* chore: bump semconvgen from 0.4.1 -> 0.11.1

I couldn't find any huge changes from the release notes; it seems the
generated wording was improved in some cases. Regardless, let us not
drift too far behind.

* feat!: update for semantic conventions @ v1.10.0

The most notable thing here is that the delta between 1.8.0 and 1.10.0
results in a few fields going away. This will be a breaking change for
anyone who happens to be depending upon these fields:

- `DB_CASSANDRA_KEYSPACE`
- `DB_HBASE_NAMESPACE`

Users should instead use `DB_NAME`, according to
[open-telemetry/opentelemetry-specification#1973](https://github.com/open-telemetry/opentelemetry-specification/pull/1973).

Otherwise, we get the standard deal:

- A plethora of new, exciting fields!
- Reworded explanatory text!

Overall, not a terribly exciting change, but we must often take joy in
the small things, and I choose to do so here!

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>