feat!: Bump required ruby version to 2.6 across everything
2.5 and 2.6 are officially EOL, so we bump the required version.

You might be asking "if 2.6 is EOL, why aren't we bumping to 2.7"?

The answer is because we are dropping official support for _MRI_ 2.6,
but we will still make a best-effort attempt to support it for JRuby.
That project has not reached 2.7 compatibility yet and we feel it's
still worth trying to maintain for that community.

However, it is indeed best-effort support, and while this constraint
means that MRI 2.6 will be able to _install_ these gems, it does not
imply _support_ for them on that runtime.