Drop support for MRI Ruby 2.5 and 2.6 (EOL) (#1218)
* ci: Stop testing 2.5 and 2.6

They are EOL.

* docs: Update references to old ruby versions in our docs

We use Ruby 3.1 in our bug report template, and reference the 2.7.6
version of docs elsewhere. There is no particular rationale as to why I
used two different versions here.

* build: Bump target rubocop version

We are dropping support for MRI 2.5 and 2.6, but maintaining support for
jruby (which is only compatible with 2.6 at this time).

So, for now, we bump our target rubocop version to 2.6.

* feat!: Bump required ruby version to 2.6 across everything

2.5 and 2.6 are officially EOL, so we bump the required version.

You might be asking "if 2.6 is EOL, why aren't we bumping to 2.7"?

The answer is because we are dropping official support for _MRI_ 2.6,
but we will still make a best-effort attempt to support it for JRuby.
That project has not reached 2.7 compatibility yet and we feel it's
still worth trying to maintain for that community.

However, it is indeed best-effort support, and while this constraint
means that MRI 2.6 will be able to _install_ these gems, it does not
imply _support_ for them on that runtime.

* feat!: Remove `Random.bytes` compatibility layer

Now that we have removed support for Ruby 2.5 and 2.6, we can remove the
indirection layer we kept for `Random.bytes`. This helps to future-proof
us as well, since `Random::DEFAULT` is deprecated in Ruby 3.0+.

* docs: Clarify support policy around 2.6.8 and jruby

Per a conversation in slack, we still wish to deprecate support for MRI
ruby `2.6`, even though we must technically maintain compatibility if we
wish to support jruby.

We will provide jruby support on a best-effort basis.

Co-authored-by: Robert <robertlaurin@users.noreply.github.com>